Add new role in between technical admin and country manager
git-svn-id: https://build.archivesportaleurope.net/svn/apex/branches/backend-project-2.0@4504 9fd842fd-667d-44da-918d-ff91420b3ae1

+++ Dashboard/src/main/java/eu/apenet/dashboard/security/actions/SessionManagementAction.java
@@ -1,6 +1,7 @@
 package eu.apenet.dashboard.security.actions;
 
 import eu.apenet.dashboard.AbstractAction;
+import eu.apenet.dashboard.security.SecurityContext;
 import eu.apenet.dashboard.security.SecurityContextContainer;
 import eu.apenet.dashboard.security.SecurityService;
 
@@ -30,7 +31,8 @@ public class SessionManagementAction extends AbstractAction {
 		return SUCCESS;
 	}
 	public String deleteSession(){
-		SecurityService.deleteSession(sessionId);
+		if(SecurityContext.get().isAdmin())
+			SecurityService.deleteSession(sessionId);
 		return SUCCESS;
 	}
